refactor: move experiments and tests into a "temp" folder
