+++ packages/client/hmi-client/src/temp/README.md
@@ -1,0 +1,1 @@
+These are engineering/testing artifacts used for proof-of-concepts that leverage some of the application's technologies, but they are NOT a part of the application proper.
