+++ packages/client/hmi-client/src/temp/TA2Playground.vue
@@ -6,7 +6,6 @@ import _ from 'lodash';
 import { defineComponent, ref } from 'vue';
 import { fetchStratificationResult } from '@/services/models/stratification-service';
 import { runDagreLayout, D3SelectionINode, D3SelectionIEdge } from '@/services/graph';
-import { logger } from '@/utils/logger';
 import API from '@/api/api';
 
 interface NodeData {
@@ -203,7 +202,7 @@ let numWolves = 10;
 export default defineComponent({
 	name: 'TA2Playground',
 	async mounted() {
-		logger.info('TA2 Playground initialized');
+		console.log('TA2 Playground initialized');
 
 		const playground = document.getElementById('playground') as HTMLDivElement;
 		renderer = new SampleRenderer({
@@ -394,7 +393,7 @@ export default defineComponent({
 		async jsonOutput() {
 			const resp = await API.get(`model-service/models/${modelId}/json`);
 			const output = await resp.data;
-			logger.info(output);
+			console.log(output);
 
 			if (petriNetValidator(output) === true) {
 				modelA = output;
@@ -435,7 +434,7 @@ export default defineComponent({
 			this.jsonOutput();
 		},
 		async addPlace() {
-			logger.info('add place');
+			console.log('add place');
 			placeCounter++;
 			const id = `p-${placeCounter}`;
 
@@ -462,7 +461,7 @@ export default defineComponent({
 			this.jsonOutput();
 		},
 		async addTransition() {
-			logger.info('add transition');
+			console.log('add transition');
 			transitionCounter++;
 			const id = `t-${transitionCounter}`;
 
@@ -558,7 +557,7 @@ export default defineComponent({
 				this.stateNamesA.length < 1 ||
 				this.stateNamesB.length < 1
 			) {
-				logger.info('Not enough states');
+				console.log('Not enough states');
 				return;
 			}
 
@@ -570,14 +569,14 @@ export default defineComponent({
 					modelB: stateNamesArrayB[i].trim()
 				});
 			}
-			logger.info(modelA);
+			console.log(modelA);
 			const resp = await API.post(`model-service/models/model-composition`, {
 				modelA,
 				modelB,
 				statesToMerge
 			});
 			mergedModel = await resp.data;
-			logger.info(`Merged petrinet ${mergedModel}`);
+			console.log(`Merged petrinet ${mergedModel}`);
 
 			g3 = parsePetriNet2IGraph(mergedModel);
 			g3 = runDagreLayout(_.cloneDeep(g3));
@@ -600,7 +599,7 @@ export default defineComponent({
 
 				const modelData = await resp.data;
 				modelId = modelData.id;
-				logger.info(`Model ID: ${modelId}`); // currently required for testing
+				console.log(`Model ID: ${modelId}`); // currently required for testing
 
 				let modelServiceNodes: { name: string; type: string }[] = [];
 				let modelServiceEdges: { source: string; target: string }[] = [];
@@ -644,15 +643,15 @@ export default defineComponent({
 			this.jsonOutput();
 		},
 		async stratify() {
-			logger.info('Start stratify');
+			console.log('Start stratify');
 			try {
 				const outputModel = await fetchStratificationResult(
 					this.stratifyModelA,
 					this.stratifyModelB,
 					this.stratifyTypeModel
 				);
-				logger.info('Result');
-				logger.info(outputModel);
+				console.log('Result');
+				console.log(outputModel);
 				this.createModel(outputModel, true);
 			} catch (e: any) {
 				logger.error(e.message);
