+++ packages/client/hmi-client/src/temp/theia.vue
